Clear request script
