--- substrate/scripts/stress/assets/types.json
@@ -1,8 +1,0 @@
-{
-    "Address": "AccountId",
-    "LookupSource": "AccountId",
-    "String": "Vec<u8>",
-    "TxMessage": {
-        "tx": "String"
-    }
-}
