--- substrate/scripts/stress/src/scenarios/index.js
@@ -1,1 +1,0 @@
-module.exports.flow = require('./flow');
